another WIP checkpoint. * setuid() added to run child procs * checks to see if acitons can be run * tidying
