+++ cmd/geneosctl/netprobe.go
@@ -29,6 +29,9 @@ func NewNetprobe(name string) (c *NetprobeComponent) {
 
 func netprobeCmd(c Component) (args, env []string) {
 	logFile := filepath.Join(getString(c, Prefix(c)+"LogD"), getString(c, Prefix(c)+"LogF"))
+	args = []string{
+		Name(c),
+	}
 	env = append(env, "LOGFILE="+logFile)
 	return
 }
